requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
python-dotenv
spotipy
itsdangerous
ytmusicapi
firebase-admin
google-auth-oauthlib
psycopg2-binary